@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "final-form": "^4.20.1",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-final-form": "^6.5.1",
-    "react-scripts": "3.4.3",
-    "uuid": "^8.3.0"
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "final-form": "^4.20.10",
+    "prop-types": "^15.8.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-final-form": "^6.5.9",
+    "react-scripts": "5.0.1",
+    "uuid": "^11.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
